5.6.0.60100
- New slash command "/titan silent" will toggle the Silent Load option
for Titan Panel, removing the clutter from PLAYER_ENTERING_WORLD event
in chat frame if it is desired. (Issue #863)
- **TitanBag**: Added displays for all profession bags. If "Ignore
Profession Bags" is selected, this actually happens now. If colored text
is selected, each profession has a color associated with it on the bar.
(Issue #895)
- Many updated to all localization files to deal with the TitanBag
changes listed above.
- Replaced TITAN_COORDS entries within all localizations to
TITAN_LOCATION.
- Removed TITAN_MONEY_FORMAT from all localizations as it is no longer
used.