5.6.21.60200
- **TitanLootType**: Still trying to get the recent changes since
Pandaria and WoD incorporated into this plugin. The process has not been
simple as there is a lack of good global data available on-line and I'm
having to mine the data myself with /dump commands and manual diffs on
old files. There will be more updates to this plugin on a regular basis
until it is complete.