5.9.0.70000
You can now enable the OrderHallCommandBar if you wish. All you need to
do is to type "/titan orderhall" without the quotes and it will show the
OrderHallCommandBar. Please note, it will require a /reload for it to
appear. The default behavior is to hide the OrderHallCommandBar so if
you don't want to see it, you don't need to do anything. One more item,
I have not been able to figure out how to move the OrderHallCommandBar
frame as it doesn't exist until you enter the Order Hall. As such, I've
been unable to embed a shift for the top bars when it appears. If
someone has worked this out, please let me know because I've looked in
other packages and they consistently just disable the
OrderHallCommandBar to avoid this issue. I'd like it to be a bit more
elegant but the solution evades me.