5.9.6.70100
Curse has gone insane again. They made a major site modification right
before a long national holiday. Files are being modified before being
approved. I uploaded 5.9.5 on Friday and it was posted on Wednesday.
Also, the file size went from 677K down to 650K. People are confused.
I'm confused. Curse is confused based on what I've seen in the
CurseForge forums so this is a rerelease of 5.9.5 and I hope it's not
truncated.