5.12.0.70200
Removed the !NoTaint_UIDropDownMenu library from the package and
replaced it with a !LibUIDropDownMenu which is being maintained by
Arith, the maintainer of Atlas. This should prevent the library
versioning errors we were seeing that were throwing all of the errors in
unrelated addons. It is safe to remove the !NoTaint_UIDropDownMenu
folder from your Titan/libs folder.